requirements.txt
# Core data analysis libraries
pandas>=2.1.0
numpy>=1.25.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.2

# Document processing
docx2txt>=0.8
pypdf>=3.17.0

# Machine learning and statistics
scikit-learn>=1.3.0
statsmodels>=0.14.0

# File format support
pyarrow>=14.0.0
PyYAML>=6.0.1

# Text encoding detection
chardet>=5.2.0

# Environment variables
python-dotenv>=1.0.0

# AI/LLM integration
groq>=0.4.1

# Web application framework
streamlit>=1.28.0

# Image processing
Pillow>=10.1.0

# Additional utilities
openpyxl>=3.1.0
xlrd>=2.0.1

# Built-in modules (no installation needed):
# - os, io, sqlite3, json, xml.etree.ElementTree, re, typing